+++ src/pages/Launchpad/index.tsx
@@ -4,7 +4,7 @@ import Fees from "./Fees";
 import Layout from "./Layout";
 import Management from "./Management";
 import Maturity from "./Maturity";
-import NetworkOption from "./NetworkOption";
+import New from "./New";
 import Splits from "./Splits";
 import Start from "./Start";
 import Success from "./Success";
@@ -34,10 +34,8 @@ export default function Launchpad() {
         element={<Success classes="justify-self-center my-10 sm:my-20 " />}
       />
       <Route
-        path={routes.network}
-        element={
-          <NetworkOption classes="justify-self-center my-10 sm:my-20 " />
-        }
+        path={routes.new}
+        element={<New classes="justify-self-center my-10 sm:my-20 " />}
       />
       <Route
         index
